+++ snapshot/sink_test.go
@@ -65,6 +65,9 @@ func Test_NewSinkOpenCloseFail(t *testing.T) {
 	}
 }
 
+// Test_SinkFullSnapshot tests that multiple full snapshots are
+// written to the Store correctly. The closing of files is awkward
+// on Windows, so this test is a little more involved.
 func Test_SinkFullSnapshot(t *testing.T) {
 	store := mustStore(t)
 	sink := NewSink(store, makeRaftMeta("snap-1234", 3, 2, 1))
@@ -81,6 +84,7 @@ func Test_SinkFullSnapshot(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Failed to copy SQLite file: %v", err)
 	}
+	sqliteFile.Close() // Reaping will fail on Windows if file is not closed.
 	if n != mustGetFileSize(t, "testdata/db-and-wals/backup.db") {
 		t.Fatalf("Unexpected number of bytes copied: %d", n)
 	}
@@ -107,6 +111,7 @@ func Test_SinkFullSnapshot(t *testing.T) {
 	if !compareReaderToFile(t, fd, "testdata/db-and-wals/backup.db") {
 		t.Fatalf("Snapshot data does not match")
 	}
+	fd.Close()
 
 	if fn, err := store.FullNeeded(); err != nil {
 		t.Fatalf("Failed to check if full snapshot needed: %v", err)
@@ -128,6 +133,7 @@ func Test_SinkFullSnapshot(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Failed to copy second SQLite file: %v", err)
 	}
+	sqliteFile2.Close()
 	if n != mustGetFileSize(t, "testdata/db-and-wals/full2.db") {
 		t.Fatalf("Unexpected number of bytes copied: %d", n)
 	}
@@ -154,6 +160,7 @@ func Test_SinkFullSnapshot(t *testing.T) {
 	if !compareReaderToFile(t, fd2, "testdata/db-and-wals/full2.db") {
 		t.Fatalf("second full snapshot data does not match")
 	}
+	fd2.Close()
 
 	// Check that setting FullNeeded flag works.
 	if fn, err := store.FullNeeded(); err != nil {
@@ -186,6 +193,7 @@ func Test_SinkFullSnapshot(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Failed to copy second SQLite file: %v", err)
 	}
+	sqliteFile3.Close()
 	if err := sink.Close(); err != nil {
 		t.Fatalf("Failed to close sink: %v", err)
 	}
@@ -194,6 +202,33 @@ func Test_SinkFullSnapshot(t *testing.T) {
 	} else if fn {
 		t.Errorf("Expected full snapshot not to be needed, but it is")
 	}
+
+	// Make sure Store returns correct snapshot.
+	expMeta3 := makeRaftMeta("snap-91011", 5, 4, 3)
+	metas3, err := store.List()
+	if err != nil {
+		t.Fatalf("Failed to list snapshots: %v", err)
+	}
+	if len(metas3) != 1 {
+		t.Fatalf("Expected 1 snapshot, got %d", len(metas))
+	}
+	compareMetas(t, expMeta3, metas3[0])
+
+	// Look inside store, make sure everything was reaped correctly.
+	files, err := os.ReadDir(store.Dir())
+	if err != nil {
+		t.Fatalf("Failed to read dir: %v", err)
+	}
+	if len(files) != 2 {
+		t.Fatalf("Expected 2 files, got %d, %s", len(files), files)
+	}
+	if !fileExists(filepath.Join(store.Dir(), "snap-91011.db")) {
+		t.Fatalf("Latest snapshot SQLite file does not exist")
+	}
+	if !dirExists(filepath.Join(store.Dir(), "snap-91011")) {
+		t.Fatalf("Latest snapshot directory does not exist")
+	}
+
 }
 
 // Test_SinkWALSnapshotEmptyStoreFail ensures that if a WAL file is