+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 8.16.3 (unreleased)
 ### Implementation changes and bug fixes
 - [PR #1582](https://github.com/rqlite/rqlite/pull/1582): Explicitly switch any upgraded v7 Snapshot to WAL mode.
+- [PR #1583](https://github.com/rqlite/rqlite/pull/1583): Loading a database invalidates existing snapshots.
 
 ## 8.16.2 (January 9th 2024)
 This releases fixes a bug related to Raft snapshot management. While it's an edge case, and can only happen if rqlited is hard-killed at a very specific point, all 8.x users should upgrade to this release.